+++ src/validataclass/helpers/__init__.py
@@ -5,6 +5,7 @@ Use of this source code is governed by an MIT-style license that can be found in
 """
 
 from .dataclass_defaults import Default, DefaultFactory, DefaultUnset, NoDefault
+from .dataclass_mixins import ValidataclassMixin
 from .dataclasses import validataclass, validataclass_field
 from .datetime_range import BaseDateTimeRange, DateTimeRange, DateTimeOffsetRange
 from .unset_value import UnsetValue, UnsetValueType, OptionalUnset, OptionalUnsetNone